Update scripts, add crowdin:sync command, update main.yml to deploy localized versions
