+++ crowdin.yml
@@ -4,10 +4,9 @@ preserve_hierarchy: true
 files:
   # JSON translation files
   - source: /i18n/en/**/*
-    translation: "/i18n/%two_letters_code%/**/%original_file_name%"
+    translation: "/i18n/%two_letters_code%/%original_path%/%original_file_name%"
   # Docs Markdown files
   - source: /docs/**/*
-    translation: /i18n/%two_letters_code%/docusaurus-plugin-content-docs/current/**/%original_file_name%
-    skip_untranslated_files: false
+    translation: /i18n/%two_letters_code%/docusaurus-plugin-content-docs/current/%original_path%/%original_file_name%
   - source: /versioned_docs/**/*
-    translation: /i18n/%two_letters_code%/docusaurus-plugin-content-docs/**/%original_file_name%
+    translation: /i18n/%two_letters_code%/docusaurus-plugin-content-docs/%original_path%/%original_file_name%
