+++ package.json
@@ -11,12 +11,12 @@
     "clear": "docusaurus clear",
     "serve": "docusaurus serve",
     "crowdin": "crowdin",
-    "crowdin:sync": "docusaurus write-translations && crowdin upload && crowdin download",
+    "crowdin:sync": "docusaurus write-translations --locale en && crowdin upload && crowdin download && node .github/scripts/reorganize.js",
     "copy-all-versions": "node copy-all-versions.js",
     "version-update": "node .github/scripts/version-update",
     "write-translations": "docusaurus write-translations",
-    "write-all-translations": "node write-all-translations.js",
     "write-heading-ids": "docusaurus write-heading-ids",
+    "reorganize": "node .github/scripts/reorganize.js",
     "typecheck": "tsc"
   },
   "dependencies": {
